+++ R/terradish-package.R
@@ -11,12 +11,12 @@
 #' @importFrom Rcpp evalCpp
 #' @importFrom MASS ginv
 #' @importFrom Matrix Cholesky Diagonal forceSymmetric rowSums solve sparseMatrix t update
-#' @importFrom methods new setRefClass
+#' @importFrom methods as new setRefClass
 #' @importFrom multiScaleR kernel_scale.raster
 #' @importFrom nlme gls
 #' @importFrom parallel clusterEvalQ clusterExport makeCluster parLapply stopCluster
 #' @importFrom stats AIC D anova as.dist as.formula coef cov2cor delete.response dist
-#' @importFrom stats dnorm fft fitted lm logLik model.matrix optimize pchisq plogis
+#' @importFrom stats dnorm fft fitted formula lm logLik model.matrix optim optimize pchisq plogis
 #' @importFrom stats pnorm printCoefmat prcomp qlogis qnorm reformulate residuals rnorm
 #' @importFrom stats rWishart sd setNames sigma simulate terms
 #' @importFrom terra adjacent cellFromXY crop geom global is.factor levels ncell nlyr patches rast